+++ src/main/java/org/sciborgs1155/robot/intake/NoIntake.java
@@ -1,0 +1,17 @@
+package org.sciborgs1155.robot.intake;
+
+import monologue.Annotations.Log;
+
+public class NoIntake implements IntakeIO {
+  @Override
+  public void setPower(double percentage) {}
+
+  @Override
+  @Log.NT
+  public boolean beambreak() {
+    return false;
+  }
+
+  @Override
+  public void close() {}
+}
